Add endpoint authorization

+++ task/src/main/java/com/project/taskmanager/task/context/comment/CommentEndpoint.java
@@ -2,6 +2,7 @@ package com.project.taskmanager.task.context.comment;
 
 import com.project.taskmanager.task.context.comment.dto.CommentRequest;
 import com.project.taskmanager.task.context.comment.dto.CommentResponse;
+import com.project.taskmanager.task.infrastructure.UserUtility;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -12,6 +13,7 @@ import java.util.List;
 @RequestMapping("/api/comment")
 @RequiredArgsConstructor
 public class CommentEndpoint {
+    private final UserUtility userUtility;
     private final CommentService commentService;
 
     @PostMapping("/create/{taskId}")
@@ -21,11 +23,15 @@ public class CommentEndpoint {
 
     @GetMapping("/{id}")
     public ResponseEntity<CommentResponse> getCommentById(@PathVariable Long id) {
+        if (!userUtility.isAdmin())
+            return ResponseEntity.badRequest().build();
         return ResponseEntity.ok(commentService.getCommentById(id));
     }
 
     @GetMapping("/all")
     public ResponseEntity<List<CommentResponse>> getAllComments() {
+        if (!userUtility.isAdmin())
+            return ResponseEntity.badRequest().build();
         return ResponseEntity.ok(commentService.getAllComments());
     }
 
@@ -36,11 +42,15 @@ public class CommentEndpoint {
 
     @PatchMapping("/update/{id}")
     public ResponseEntity<CommentResponse> updateComment(@PathVariable Long id, @RequestBody CommentRequest request) {
+        if (!userUtility.isAdmin() && !userUtility.isCommentCreator(id))
+            return ResponseEntity.badRequest().build();
         return ResponseEntity.ok(commentService.updateComment(id, request));
     }
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<String> deleteComment(@PathVariable Long id) {
+        if (!userUtility.isAdmin() && !userUtility.isCommentCreator(id))
+            return ResponseEntity.badRequest().build();
         commentService.deleteComment(id);
         return ResponseEntity.ok("Comment deleted successfully");
     }
