+++ task/src/main/java/com/project/taskmanager/task/context/task/TaskEndpoint.java
@@ -2,8 +2,10 @@ package com.project.taskmanager.task.context.task;
 
 import com.project.taskmanager.task.context.task.dto.TaskRequest;
 import com.project.taskmanager.task.context.task.dto.TaskResponse;
+import com.project.taskmanager.task.infrastructure.UserUtility;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -12,6 +14,7 @@ import java.util.List;
 @RequestMapping("/api/task")
 @RequiredArgsConstructor
 public class TaskEndpoint {
+    private final UserUtility userUtility;
     private final TaskService taskService;
 
     @PostMapping("/create")
@@ -31,11 +34,17 @@ public class TaskEndpoint {
 
     @PatchMapping("/update/{id}")
     public ResponseEntity<TaskResponse> updateTask(@PathVariable Long id, @RequestBody TaskRequest request) {
+        if (!userUtility.isAdmin() && !userUtility.isTaskCreator(id)) {
+            return ResponseEntity.badRequest().build();
+        }
         return ResponseEntity.ok(taskService.updateTask(id, request));
     }
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<String> deleteTask(@PathVariable Long id) {
+        if (!userUtility.isAdmin() && !userUtility.isTaskCreator(id)) {
+            return ResponseEntity.badRequest().build();
+        }
         taskService.deleteTask(id);
         return ResponseEntity.ok("Task deleted successfully");
     }
